+++ R/emulate.R
@@ -72,154 +72,116 @@ emulate.telemetry <- function(object,CTMM,fast=FALSE,...)
 
 
 #####################################
-# fast here debiases for fast emulate
-ctmm.boot <- function(data,CTMM,fast=FALSE,n=1E4,mc.cores=NULL,...)
+# multi-estimator parametric bootstrap
+# + concurrent double-bootstrap AICc
+ctmm.boot <- function(data,CTMM,method=c('HREML','pREML','pHREML'),error=0.01,mc.cores=NULL,AICc=FALSE,...)
 {
   if(is.null(mc.cores)) { mc.cores <- detectCores() } # Windows safe wrapper
 
-  NAMES <- dimnames(CTMM$COV)[[1]] # parameters to extract & debias
+  # toss location information for simulations unconditioned
+  FRAME <- data
+  FRAME[,CTMM$axes] <- NULL
 
-  Replicate <- function(i)
-  {
-    FIT <- emulate(CTMM,data=data,fast=FALSE,...)
-    FIT <- get.parameters(FIT,NAMES)
-    return(FIT)
-  }
-
-  PARS <- mclapply(1:n,Replicate,mc.cores=mc.cores)
-  PARS <- simplify2array(PARS)
-  rownames(PARS) <- NAMES
-
-  sigma <- attr(CTMM$sigma,"par")
-
-  # create MSD / average variance data row
-  if(CTMM$isotropic)
-  { STAT <- PARS['area',] }
-  else
-  { STAT <- PARS['area',] * cosh(PARS['eccentricity',]/2) }
-  PARS <- rbind(PARS,STAT)
-  rownames(PARS)[nrow(PARS)] <- 'var'
-
-  ###################################
-  ### debias important quantities ###
-
-  # debias p1 given p1 -> p2
-  # p1: null model parameter
-  # p2 = p1 + bias
-  # lower if p>0
-  debias <- function(p1,p2,lower=FALSE)
-  {
-    if(lower && p1<p2) # relative/scale debias formula (same to first order)
-    { p0 <- p1^2 / p2 } # remember sign for magnitudes...
-    else # absolute/shift debias formula (exact on first order)
-    { p0 <- 2*p1 - p2 }
-
-    return(p0)
-  }
-
-  ### (SCALE) standard area debias
-  STAT <- mean(PARS['area',])
-  sigma['area'] <- debias(sigma['area'],STAT,lower=TRUE)
-
-  if(!CTMM$isotropic)
-  {
-    ### (SCALE) average variance debias
-    STAT <- mean( PARS['var',] )
-    # cosh factor
-    STAT <- STAT/sigma['area']
-    # (SCALE) debiased via cosh factor
-    STAT <- debias(cosh(sigma['eccentricity']/2),STAT,lower=TRUE)
-    if(STAT>1)
-    {
-      sigma['eccentricity'] <- 2*acosh(STAT)
-
-      # (SHIFT) debiased angle
-      sigma['angle'] <- debias(sigma['angle'],mean(PARS['angle',]))
-    }
-    else # eccentricity is explained entirely by bias
-    {
-      sigma['eccentricity'] <- 0
-      sigma['angle'] <- 0
-      CTMM$isotropic <- FALSE
-    }
-  }
-
-  # set sigma parameters
-  CTMM$sigma <- covm(sigma,isotropic=CTMM$isotropic)
-  VAR <- sigma['area'] * cosh(sigma['eccentricity']/2)
-
-  ### (SCALE) average diffusion rate debias
-  if('tau position' %in% NAMES)
-  {
-    # average diffusion rate
-    STAT <- mean( PARS['var',] / PARS['tau position',] )
-    # tau position
-    STAT <- VAR/STAT
-    # (SCALE) dibias via tau
-    CTMM$tau[1] <- debias(CTMM$tau[1],STAT,lower=TRUE)
-  }
-
-  ### (SCALE) MS velocity debias
-  if('tau velocity' %in% NAMES)
-  {
-    if(CTMM$range)
-    {
-      STAT <- mean( PARS['var',] / PARS['tau position',] / PARS['tau velocity',] )
-      STAT <- VAR / CTMM$tau[1] / STAT
-    }
-    else
-    {
-      STAT <- mean( PARS['var',] / PARS['tau velocity',] )
-      STAT <- VAR / STAT
-    }
-    # debias via tau velocity
-    CTMM$tau[2] <- debias(CTMM$tau[2],STAT,lower=TRUE)
-  }
-
-  ### (SCALE) MS velocity debias
-  if('circle' %in% NAMES)
-  {
-    SIGN <- sign(CTMM$circle)
-    STAT <- mean( PARS['var',] * PARS['circle',]^2 )
-    # abs(circle)
-    STAT <- sqrt( STAT / VAR ) * SIGN
-    CTMM$circle <- debias(CTMM$circle,STAT)
-    if(sign(CTMM$circle)!=SIGN) { CTMM$circle <- FALSE } # circle is entirely explained by bias
-  }
-
-  ### (SCALE) error
-  if('error' %in% NAMES)
-  {
-    STAT <- mean( PARS['error',] )
-    CTMM$error <- debias(CTMM$error,STAT,lower=TRUE)
-  }
-
-  ### parameter estimate covariances ###
-
-  P0 <- get.parameters(CTMM,NAMES)
-  PARS <- PARS[NAMES,] # discard derived parameters
+  # parameters to extract & debias
+  NAMES <- dimnames(CTMM$COV)[[1]]
+  DIM <- length(NAMES) * length(method)
 
   # positive variables (potential)
   POS <- c("area","tau position","tau velocity","error")
   # included variables that are positive
   POS <- POS[POS %in% NAMES]
+
   # log transform positive parameters
-  if(length(POS)) { PARS[POS,] <- log(PARS[POS,]) }
-
-  COV <- stats::cov(t(PARS))
-
-  # log transform back
-  if(length(POS))
+  Transform <- function(p,inverse=FALSE)
   {
-    #P0 <- mean(t(PARS[POS,]))
+    if(length(POS))
+    {
+      if(inverse) { p[POS,] <- exp(p[POS,]) }
+      else { p[POS,] <- log(p[POS,]) }
+    }
 
-    COV[POS,] <- P0[POS] * COV[POS,]
-    COV[,POS] <- t(P0[POS] * t(COV[,POS]))
+    return(p)
   }
 
-  # recalculate trend terms
-  CTMM <- ctmm.loglike(data,CTMM,verbose=TRUE,profile=FALSE)
+  # simulate with errors & return parameter estimates
+  Replicate <- function(i,DATA=simulate(CTMM,data=FRAME))
+  {
+    # apply all estimators, recycling previous fit as first guess
+    FIT <- list(CTMM=CTMM) # first guess (to be removed)
+    for(m in method) { FIT <- c(FIT,ctmm.fit(FRAME,FIT[[length(FIT)]],method=m,COV=FALSE,...)) }
+    FIT <- FIT[-1] # remove first (fake) guess
 
+    # extract parameters
+    FIT <- sapply(FIT,function(f){ get.parameters(f,NAMES) }) # [parameter,method]
+    rownames(FIT) <- NAMES
+    colnames(FIT) <- method
+
+    # log transform positive parameters
+    FIT <- Transform(FIT)
+
+    return(FIT)
+  }
+
+  # outer loop: designing the weight matrix until estimate converges
+  EST <- Replicate(0,DATA=data) # estimates to weight (ultimately)
+  P0 <- Transform(get.parameters(CTMM,NAMES)) # null model for parametric bootstrap
+  ERROR <- Inf
+  while(ERROR>=error)
+  {
+    MEAN <- S1 <- array(0,DIM)
+    COV <- S2 <- array(0,c(DIM,DIM))
+    ERROR <- Inf
+    N <- 0
+
+    # inner loop: adding to ensemble until average converges
+    while(ERROR>=error)
+    {
+      # calculate burst of mc.cores estimates
+      ADD <- mclapply(1:mc.cores,Replicate,mc.cores=mc.cores)
+      ADD <- simplify2array(ADD) # [par,method,run]
+      ADD <- array(ADD,c(DIM,mc.cores)) # [par*method,run]
+
+      # rolling mean and variance
+      # update sums
+      S1 <- S1 + rowSums(ADD)
+      S2 <- S2 + (ADD %.% t(ADD))
+      # update sample size
+      N <- N + mc.cores
+      # normalize sums
+      MEAN <- S1/N
+      COV <- (S2-N^2*outer(MEAN))/max(1,N-1) # !!! subtract mean^2
+
+      # recalculate error
+      if(N>10) # may want to increase to 100
+      {
+        BIAS <- MEAN - P0
+        ERROR <- sqrt((BIAS %*% PDsolve(COV) %*% BIAS)/N)
+      }
+    }
+
+    ## recalculate weighted estimator
+    P1 <- EST - BIAS # debiased estimates
+    M <- expm::sqrtm(PDsolve(COV))
+    P1 <- M %*% P1 # standardized esitmates
+    M <- M %*% t(array(diag(length(NAMES)),c(length(NAMES),DIM))) # LSM
+    P1 <- t(M) %*% P1
+    M <- t(M) %*% M # precision matrix
+    COV <- PDsolve(M)
+    P1 <- M %*% P1
+
+    ## recalculate error & store best model
+    BIAS <- P1 - P0 # now the change
+    ERROR <- sqrt(BIAS %*% M %*% BIAS)
+    CTMM <- set.parameters(CTMM,Transform(P1,inverse=TRUE))
+    P1 -> P0
+  }
+
+  # transform COV
+  if(length(POS))
+  {
+    COV[POS,] <- COV[POS,]/P0[POS]
+    COV[,POS] <- COV[,POS]/P0[POS]
+  }
   CTMM$COV <- COV
 
   return(CTMM)